KAD-83: move `LambdaExceptionUtil` + tests to common package
