+++ common/kadai-common/src/main/java/io/kadai/common/internal/util/LambdaExceptionUtil.java
@@ -1,4 +1,4 @@
-package io.kadai.common.rest.util;
+package io.kadai.common.internal.util;
 
 import java.util.function.Consumer;
 import java.util.function.Function;
